requirements.txt
python-dotenv
discord[voice]